Move impl blocks to lower the number of cfg's.
